license_api/requirements.txt
# license_api/requirements.txt
functions-framework
google-cloud-firestore